Introduce worker managers package. Fixes #659
