+++ aimmo-game/simulation/worker_managers/__init__.py
@@ -1,0 +1,7 @@
+from kubernetes_worker_manager import KubernetesWorkerManager
+from local_worker_manager import LocalWorkerManager
+
+WORKER_MANAGERS = {
+    'local': LocalWorkerManager,
+    'kubernetes': KubernetesWorkerManager,
+}
